refactor: 프로필 관련 활동 서비스, 컨트롤러 분리
