+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
 
   datasource:
     url: ${DB_URL}
-    username: root
+    username: admin
     password: ${DB_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
   jpa:
